- fixed compilation warnings introduced recently
src/console/c_cmds.cpp:690:11: warning: format string is not a string literal (potentially insecure) [-Wformat-security]
src/scripting/vmthunks.cpp:2698:21: warning: comparison of integers of different signs: 'int' and 'unsigned long' [-Wsign-compare]

+++ src/scripting/vmthunks.cpp
@@ -2695,7 +2695,7 @@ static FString CheckStrfString(FString timeForm)
 		if (t.Len() == 0) return FString("%");
 
 		// Single Character
-		for (int i = 0; i < sizeof(validSingles)/sizeof(validSingles[0]); i++)
+		for (size_t i = 0; i < sizeof(validSingles)/sizeof(validSingles[0]); i++)
 		{
 			if (t[0] == validSingles[i])
 			{
